requirements.txt
fastapi
uvicorn
boto3
networkx
python-dotenv
